Metadata templates (#210)
* Allow edit templates in metadata

Allow the provider to pass in editing templates for JSAPI 4.x to work correctly for editing.

* Don't force Int Object ID

Don't force OBJECTID to be an integer if it is not a number (a GUID for instance).

* Update CHANGELOG.md

* Add Test Cases

Add test cases for string Object IDs and for templates being passed in through metadata for a layer